Several changes were made to the original script as listed below.
(1) Renamed classes and variables to be more descriptive and conforming to PEP8, as enforced by the default config of pylint.

(2) No more global variables

(3) Removed get_mask and set_mask methods so as not to change the state of Obs object. Now, mask can be either set for all variables while creating an Obs object, or set for individual variables in the get_var call.

(4) Not storing the data read from file into the Obs object. That way (a) we do not modify the state of the object and (b) we do not run the risk of this object getting large.

(5) Instead of opening the file in __init__ (and possibly closing it in __del__), files are now opened (and data read) inside a “with” block.

(6) Added the example as a unit test (das_tools/python_tools/utest/test_ncdiag.py)

(7) Not using ncdiag_functions module any more. Added functionalities from this module as static methods of Obs class.

TODO: Re-implement generating statistics